Make ClientToolManager persistent.
This commit allows a ClientToolManager instance to survive a disconnect
and reset itself, so when reconnecting the client to another server,
there is no need to delete and reconstruct the ClientToolManager.

To achieve this, this commit moves the connectionEstablished signal from
Client to Endpoint (for linking reasons). I also added an emit to
Server, just to not have misleading API.